+++ lingotrace/core/capabilities.py
@@ -11,6 +11,7 @@ PUBLIC_CAPABILITY_IDS = {
     "listening_notes",
     "source_notes",
     "review_materials",
+    "review_queue",
     "speaking_cards",
     "review_rollover",
     "total_training_dashboard",